requirements.txt
numpy-1.23.1
opencv-python-4.6.0.66
tensorflow-2.9.1
dlib
